[NEW-50407 ] Data Bite > update packages and components used (#2160)
* Data Bite > update packages and components used

* remove react-beautiful-dnd

* regenerate lock file
