+++ packages/data-bite/src/constants.ts
@@ -1,0 +1,66 @@
+export const DATA_FUNCTION_MAX = 'Max'
+export const DATA_FUNCTION_COUNT = 'Count'
+export const DATA_FUNCTION_MEAN = 'Mean (Average)'
+export const DATA_FUNCTION_MEDIAN = 'Median'
+export const DATA_FUNCTION_MIN = 'Min'
+export const DATA_FUNCTION_MODE = 'Mode'
+export const DATA_FUNCTION_RANGE = 'Range'
+export const DATA_FUNCTION_SUM = 'Sum'
+export const DATA_FUNCTIONS = [
+  DATA_FUNCTION_COUNT,
+  DATA_FUNCTION_MAX,
+  DATA_FUNCTION_MEAN,
+  DATA_FUNCTION_MEDIAN,
+  DATA_FUNCTION_MIN,
+  DATA_FUNCTION_MODE,
+  DATA_FUNCTION_RANGE,
+  DATA_FUNCTION_SUM
+]
+
+export const BITE_LOCATION_TITLE = 'title'
+export const BITE_LOCATION_BODY = 'body'
+export const BITE_LOCATION_GRAPHIC = 'graphic'
+export const BITE_LOCATIONS = {
+  graphic: 'Graphic',
+  split: 'Split Graphic and Message',
+  title: 'Value above Message',
+  body: 'Value before Message',
+  end: 'Value after Message',
+  gradient: 'Gradient'
+}
+
+export const IMAGE_POSITION_LEFT = 'Left'
+export const IMAGE_POSITION_RIGHT = 'Right'
+export const IMAGE_POSITION_TOP = 'Top'
+export const IMAGE_POSITION_BOTTOM = 'Bottom'
+export const IMAGE_POSITIONS = [IMAGE_POSITION_LEFT, IMAGE_POSITION_RIGHT, IMAGE_POSITION_TOP, IMAGE_POSITION_BOTTOM]
+
+export const DATA_OPERATOR_LESS = '<'
+export const DATA_OPERATOR_GREATER = '>'
+export const DATA_OPERATOR_LESSEQUAL = '<='
+export const DATA_OPERATOR_GREATEREQUAL = '>='
+export const DATA_OPERATOR_EQUAL = '='
+export const DATA_OPERATOR_NOTEQUAL = '≠'
+
+export const DATA_OPERATORS = [
+  DATA_OPERATOR_LESS,
+  DATA_OPERATOR_GREATER,
+  DATA_OPERATOR_LESSEQUAL,
+  DATA_OPERATOR_GREATEREQUAL,
+  DATA_OPERATOR_EQUAL,
+  DATA_OPERATOR_NOTEQUAL
+]
+
+export const HEADER_COLORS = [
+  'theme-blue',
+  'theme-purple',
+  'theme-brown',
+  'theme-teal',
+  'theme-pink',
+  'theme-orange',
+  'theme-slate',
+  'theme-indigo',
+  'theme-cyan',
+  'theme-green',
+  'theme-amber'
+]
